remove unused directory & move /app/_components to /components
