--- src/archive/chat/[bookId]/layout.tsx
@@ -1,17 +1,0 @@
-// import Navbar from "@/app/_components/navbar";
-
-// export default function ChatPageLayout({
-//   children,
-// }: {
-//   children: React.ReactNode;
-// }) {
-//   return (
-//     <div>
-//       <Navbar />
-
-//       {children}
-
-//       {/* <Footer /> */}
-//     </div>
-//   );
-// }
