+++ src/components/navbar/locale-switcher.tsx
