+++ src/components/navbar/search-bar/empty-state.tsx
